+++ release/scripts/addons/io_scene_gltf2/__init__.py
@@ -43,7 +43,7 @@ bl_info = {
     'location': 'File > Import-Export',
     'description': 'Import-Export as glTF 2.0',
     'warning': '',
-    'wiki_url': "https://github.com/KhronosGroup/glTF-Blender-IO",
+    'wiki_url': "https://docs.blender.org/manual/en/dev/addons/io_gltf2.html",
     'tracker_url': "https://github.com/KhronosGroup/glTF-Blender-IO/issues/",
     'support': 'OFFICIAL',
     'category': 'Import-Export'}
@@ -82,7 +82,6 @@ class ExportGLTF2_Base:
         items=(('GENERAL', "General", "General settings"),
                ('MESHES', "Meshes", "Mesh settings"),
                ('OBJECTS', "Objects", "Object settings"),
-               ('MATERIALS', "Materials", "Material settings"),
                ('ANIMATION', "Animation", "Animation settings")),
         name="ui_tab",
         description="Export setting categories",
@@ -156,7 +155,7 @@ class ExportGLTF2_Base:
 
     export_apply: BoolProperty(
         name='Apply Modifiers',
-        description='Apply modifiers to mesh objects',
+        description='Apply modifiers (excluding Armatures) to mesh objects',
         default=False
     )
 
@@ -238,13 +237,6 @@ class ExportGLTF2_Base:
         name='Punctual Lights',
         description='Export directional, point, and spot lights. '
                     'Uses "KHR_lights_punctual" glTF extension',
-        default=False
-    )
-
-    export_texture_transform: BoolProperty(
-        name='Texture Transforms',
-        description='Export texture or UV position, rotation, and scale. '
-                    'Uses "KHR_texture_transform" glTF extension',
         default=False
     )
 
@@ -351,7 +343,6 @@ class ExportGLTF2_Base:
             export_settings['gltf_morph_tangent'] = False
 
         export_settings['gltf_lights'] = self.export_lights
-        export_settings['gltf_texture_transform'] = self.export_texture_transform
         export_settings['gltf_displacement'] = self.export_displacement
 
         export_settings['gltf_binary'] = bytearray()
@@ -389,16 +380,12 @@ class ExportGLTF2_Base:
         if self.export_normals:
             col.prop(self, 'export_tangents')
         col.prop(self, 'export_colors')
+        col.prop(self, 'export_materials')
 
     def draw_object_settings(self):
         col = self.layout.box().column()
         col.prop(self, 'export_cameras')
         col.prop(self, 'export_lights')
-
-    def draw_material_settings(self):
-        col = self.layout.box().column()
-        col.prop(self, 'export_materials')
-        col.prop(self, 'export_texture_transform')
 
     def draw_animation_settings(self):
         col = self.layout.box().column()
@@ -424,7 +411,7 @@ class ExportGLTF2_Base:
 class ExportGLTF2(bpy.types.Operator, ExportGLTF2_Base, ExportHelper):
     """Export scene as glTF 2.0 file"""
     bl_idname = 'export_scene.gltf'
-    bl_label = 'glTF 2.0 (.glb/.gltf)'
+    bl_label = 'Export glTF 2.0'
 
     filename_ext = ''
 
@@ -436,8 +423,9 @@ def menu_func_export(self, context):
 
 
 class ImportGLTF2(Operator, ImportHelper):
+    """Load a glTF 2.0 file"""
     bl_idname = 'import_scene.gltf'
-    bl_label = 'glTF 2.0 (.glb/.gltf)'
+    bl_label = 'Import glTF 2.0'
 
     filter_glob: StringProperty(default="*.glb;*.gltf", options={'HIDDEN'})
 
@@ -497,7 +485,7 @@ class ImportGLTF2(Operator, ImportHelper):
 
 
 def menu_func_import(self, context):
-    self.layout.operator(ImportGLTF2.bl_idname, text=ImportGLTF2.bl_label)
+    self.layout.operator(ImportGLTF2.bl_idname, text='glTF 2.0 (.glb/.gltf)')
 
 
 classes = (