+++ source/blender/makesrna/intern/rna_object.c
@@ -36,6 +36,8 @@
 
 #include "BLI_utildefines.h"
 
+#include "BLT_translation.h"
+
 #include "BKE_camera.h"
 #include "BKE_collection.h"
 #include "BKE_paint.h"
@@ -345,8 +347,8 @@ static void rna_Object_active_shape_update(bContext *C, PointerRNA *ptr)
 
 				DEG_id_tag_update(ob->data, 0);
 
-				EDBM_mesh_normals_update(((Mesh *)ob->data)->edit_btmesh);
-				BKE_editmesh_tessface_calc(((Mesh *)ob->data)->edit_btmesh);
+				EDBM_mesh_normals_update(((Mesh *)ob->data)->edit_mesh);
+				BKE_editmesh_tessface_calc(((Mesh *)ob->data)->edit_mesh);
 				break;
 			case OB_CURVE:
 			case OB_SURF:
@@ -551,9 +553,9 @@ static void rna_Object_dup_collection_set(PointerRNA *ptr, PointerRNA value)
 	 */
 	if (BKE_collection_has_object_recursive(grp, ob) == 0) {
 		if (ob->type == OB_EMPTY) {
-			id_us_min(&ob->dup_group->id);
-			ob->dup_group = grp;
-			id_us_plus(&ob->dup_group->id);
+			id_us_min(&ob->instance_collection->id);
+			ob->instance_collection = grp;
+			id_us_plus(&ob->instance_collection->id);
 		}
 		else {
 			BKE_report(NULL, RPT_ERROR,
@@ -787,8 +789,8 @@ static void rna_Object_active_material_index_set(PointerRNA *ptr, int value)
 	if (ob->type == OB_MESH) {
 		Mesh *me = ob->data;
 
-		if (me->edit_btmesh)
-			me->edit_btmesh->mat_nr = value;
+		if (me->edit_mesh)
+			me->edit_mesh->mat_nr = value;
 	}
 }
 
@@ -1097,11 +1099,10 @@ static char *rna_MaterialSlot_path(PointerRNA *ptr)
 	return BLI_sprintfN("material_slots[%d]", index);
 }
 
-/* why does this have to be so complicated?, can't all this crap be
- * moved to in BGE conversion function? - Campbell *
- *
- * logic from check_body_type()
- *  */
+static PointerRNA rna_Object_display_get(PointerRNA *ptr)
+{
+	return rna_pointer_inherit_refine(ptr, &RNA_ObjectDisplay, ptr->data);
+}
 
 static char *rna_ObjectDisplay_path(PointerRNA *UNUSED(ptr))
 {
@@ -2079,11 +2080,12 @@ static void rna_def_object_display(BlenderRNA *brna)
 
 	srna = RNA_def_struct(brna, "ObjectDisplay", NULL);
 	RNA_def_struct_ui_text(srna, "Object Display", "Object display settings for 3d viewport");
-	RNA_def_struct_sdna(srna, "ObjectDisplay");
+	RNA_def_struct_sdna(srna, "Object");
+	RNA_def_struct_nested(brna, srna, "Object");
 	RNA_def_struct_path_func(srna, "rna_ObjectDisplay_path");
 
 	prop = RNA_def_property(srna, "show_shadows", PROP_BOOLEAN, PROP_NONE);
-	RNA_def_property_boolean_sdna(prop, NULL, "flag", OB_SHOW_SHADOW);
+	RNA_def_property_boolean_negative_sdna(prop, NULL, "dtx", OB_DRAW_NO_SHADOW_CAST);
 	RNA_def_property_boolean_default(prop, true);
 	RNA_def_property_ui_text(prop, "Shadow", "Object cast shadows in the 3d viewport");
 	RNA_def_property_update(prop, NC_OBJECT | ND_DRAW, NULL);
@@ -2142,6 +2144,7 @@ static void rna_def_object(BlenderRNA *brna)
 	RNA_def_property_enum_items(prop, rna_enum_object_type_items);
 	RNA_def_property_clear_flag(prop, PROP_EDITABLE);
 	RNA_def_property_ui_text(prop, "Type", "Type of Object");
+	RNA_def_property_translation_context(prop, BLT_I18NCONTEXT_ID_ID);
 
 	prop = RNA_def_property(srna, "mode", PROP_ENUM, PROP_NONE);
 	RNA_def_property_enum_sdna(prop, NULL, "mode");
@@ -2617,14 +2620,14 @@ static void rna_def_object(BlenderRNA *brna)
 	RNA_def_property_update(prop, NC_OBJECT | ND_DRAW, "rna_Object_internal_update");
 
 	prop = RNA_def_property(srna, "instance_faces_scale", PROP_FLOAT, PROP_NONE);
-	RNA_def_property_float_sdna(prop, NULL, "dupfacesca");
+	RNA_def_property_float_sdna(prop, NULL, "instance_faces_scale");
 	RNA_def_property_range(prop, 0.001f, 10000.0f);
 	RNA_def_property_ui_text(prop, "Instance Faces Scale", "Scale the face instance objects");
 	RNA_def_property_update(prop, NC_OBJECT | ND_DRAW, "rna_Object_internal_update");
 
 	prop = RNA_def_property(srna, "instance_collection", PROP_POINTER, PROP_NONE);
 	RNA_def_property_struct_type(prop, "Collection");
-	RNA_def_property_pointer_sdna(prop, NULL, "dup_group");
+	RNA_def_property_pointer_sdna(prop, NULL, "instance_collection");
 	RNA_def_property_flag(prop, PROP_EDITABLE);
 	RNA_def_property_pointer_funcs(prop, NULL, "rna_Object_dup_collection_set", NULL, NULL);
 	RNA_def_property_ui_text(prop, "Instance Collection", "Instance an existing collection");
@@ -2759,8 +2762,9 @@ static void rna_def_object(BlenderRNA *brna)
 
 	/* Object Display */
 	prop = RNA_def_property(srna, "display", PROP_POINTER, PROP_NONE);
-	RNA_def_property_pointer_sdna(prop, NULL, "display");
+	RNA_def_property_flag(prop, PROP_NEVER_NULL);
 	RNA_def_property_struct_type(prop, "ObjectDisplay");
+	RNA_def_property_pointer_funcs(prop, "rna_Object_display_get", NULL, NULL, NULL);
 	RNA_def_property_ui_text(prop, "Object Display", "Object display settings for 3d viewport");
 
 	RNA_api_object(srna);