+++ source/blender/makesrna/intern/rna_space.c
@@ -74,7 +74,7 @@ const EnumPropertyItem rna_enum_space_type_items[] = {
 	{0, "", ICON_NONE, "Animation", ""},
 	//{SPACE_ACTION, "TIMELINE", ICON_TIME, "Timeline", "Timeline and playback controls (NOTE: Switch to 'Timeline' mode)"}, /* XXX */
 	{SPACE_ACTION, "DOPESHEET_EDITOR", ICON_ACTION, "Dope Sheet", "Adjust timing of keyframes"},
-	{SPACE_IPO, "GRAPH_EDITOR", ICON_GRAPH, "Graph Editor", "Edit drivers and keyframe interpolation"},
+	{SPACE_GRAPH, "GRAPH_EDITOR", ICON_GRAPH, "Graph Editor", "Edit drivers and keyframe interpolation"},
 	{SPACE_NLA, "NLA_EDITOR", ICON_NLA, "Nonlinear Animation", "Combine and layer Actions"},
 
 	/* Scripting */
@@ -94,7 +94,7 @@ const EnumPropertyItem rna_enum_space_type_items[] = {
 	{0, "", ICON_NONE, "Data", ""},
 	{ SPACE_TOOLBAR, "TOOLBAR", ICON_TOOLBAR, "Toolbar", "A toolbar. Here you can display various button menus" },
 	{SPACE_OUTLINER, "OUTLINER", ICON_OUTLINER, "Outliner", "Overview of scene graph and all available data-blocks"},
-	{SPACE_BUTS, "PROPERTIES", ICON_PROPERTIES, "Properties", "Edit properties of active object and related data-blocks"},
+	{SPACE_PROPERTIES, "PROPERTIES", ICON_PROPERTIES, "Properties", "Edit properties of active object and related data-blocks"},
 	{SPACE_FILE, "FILE_BROWSER", ICON_FILEBROWSER, "File Browser", "Browse for files and assets"},
 	{SPACE_USERPREF, "PREFERENCES", ICON_PREFERENCES, "Preferences",
 	                 "Edit persistent configuration settings"},
@@ -386,11 +386,11 @@ static StructRNA *rna_Space_refine(struct PointerRNA *ptr)
 	switch (space->spacetype) {
 		case SPACE_VIEW3D:
 			return &RNA_SpaceView3D;
-		case SPACE_IPO:
+		case SPACE_GRAPH:
 			return &RNA_SpaceGraphEditor;
 		case SPACE_OUTLINER:
 			return &RNA_SpaceOutliner;
-		case SPACE_BUTS:
+		case SPACE_PROPERTIES:
 			return &RNA_SpaceProperties;
 		case SPACE_FILE:
 			return &RNA_SpaceFileBrowser;
@@ -1218,13 +1218,13 @@ static void rna_SpaceTextEditor_updateEdited(Main *UNUSED(bmain), Scene *UNUSED(
 /* note: this function exists only to avoid id refcounting */
 static void rna_SpaceProperties_pin_id_set(PointerRNA *ptr, PointerRNA value)
 {
-	SpaceButs *sbuts = (SpaceButs *)(ptr->data);
+	SpaceProperties *sbuts = (SpaceProperties *)(ptr->data);
 	sbuts->pinid = value.data;
 }
 
 static StructRNA *rna_SpaceProperties_pin_id_typef(PointerRNA *ptr)
 {
-	SpaceButs *sbuts = (SpaceButs *)(ptr->data);
+	SpaceProperties *sbuts = (SpaceProperties *)(ptr->data);
 
 	if (sbuts->pinid)
 		return ID_code_to_RNA_type(GS(sbuts->pinid->name));
@@ -1234,7 +1234,7 @@ static StructRNA *rna_SpaceProperties_pin_id_typef(PointerRNA *ptr)
 
 static void rna_SpaceProperties_pin_id_update(Main *UNUSED(bmain), Scene *UNUSED(scene), PointerRNA *ptr)
 {
-	SpaceButs *sbuts = (SpaceButs *)(ptr->data);
+	SpaceProperties *sbuts = (SpaceProperties *)(ptr->data);
 	ID *id = sbuts->pinid;
 
 	if (id == NULL) {
@@ -1263,7 +1263,7 @@ static void rna_SpaceProperties_pin_id_update(Main *UNUSED(bmain), Scene *UNUSED
 
 static void rna_SpaceProperties_context_set(PointerRNA *ptr, int value)
 {
-	SpaceButs *sbuts = (SpaceButs *)(ptr->data);
+	SpaceProperties *sbuts = (SpaceProperties *)(ptr->data);
 
 	sbuts->mainb = value;
 	sbuts->mainbuser = value;
@@ -1273,7 +1273,7 @@ static const EnumPropertyItem *rna_SpaceProperties_context_itemf(
         bContext *UNUSED(C), PointerRNA *ptr,
         PropertyRNA *UNUSED(prop), bool *r_free)
 {
-	SpaceButs *sbuts = (SpaceButs *)(ptr->data);
+	SpaceProperties *sbuts = (SpaceProperties *)(ptr->data);
 	EnumPropertyItem *item = NULL;
 	int totitem = 0;
 
@@ -1362,7 +1362,7 @@ static const EnumPropertyItem *rna_SpaceProperties_context_itemf(
 
 static void rna_SpaceProperties_context_update(Main *UNUSED(bmain), Scene *UNUSED(scene), PointerRNA *ptr)
 {
-	SpaceButs *sbuts = (SpaceButs *)(ptr->data);
+	SpaceProperties *sbuts = (SpaceProperties *)(ptr->data);
 	/* XXX BCONTEXT_DATA is ugly, but required for lights... See T51318. */
 	if (ELEM(sbuts->mainb, BCONTEXT_WORLD, BCONTEXT_MATERIAL, BCONTEXT_TEXTURE, BCONTEXT_DATA)) {
 		sbuts->preview = 1;
@@ -1586,7 +1586,7 @@ static void rna_SpaceDopeSheetEditor_mode_update(bContext *C, PointerRNA *ptr)
 static void rna_SpaceGraphEditor_display_mode_update(bContext *C, PointerRNA *ptr)
 {
 	ScrArea *sa = rna_area_from_space(ptr);
-	SpaceIpo *sipo = (SpaceIpo *)ptr->data;
+	SpaceGraph *sipo = (SpaceGraph *)ptr->data;
 
 	/* for "Drivers" mode, enable all the necessary bits and pieces */
 	if (sipo->mode == SIPO_MODE_DRIVERS) {
@@ -1602,7 +1602,7 @@ static void rna_SpaceGraphEditor_display_mode_update(bContext *C, PointerRNA *pt
 
 static bool rna_SpaceGraphEditor_has_ghost_curves_get(PointerRNA *ptr)
 {
-	SpaceIpo *sipo = (SpaceIpo *)(ptr->data);
+	SpaceGraph *sipo = (SpaceGraph *)(ptr->data);
 	return (BLI_listbase_is_empty(&sipo->runtime.ghost_curves) == false);
 }
 
@@ -2293,7 +2293,7 @@ static void rna_def_space_outliner(BlenderRNA *brna)
 	};
 
 	srna = RNA_def_struct(brna, "SpaceOutliner", "Space");
-	RNA_def_struct_sdna(srna, "SpaceOops");
+	RNA_def_struct_sdna(srna, "SpaceOutliner");
 	RNA_def_struct_ui_text(srna, "Space Outliner", "Outliner space data");
 
 	prop = RNA_def_property(srna, "display_mode", PROP_ENUM, PROP_NONE);
@@ -2400,6 +2400,7 @@ static void rna_def_space_outliner(BlenderRNA *brna)
 	RNA_def_property_enum_sdna(prop, NULL, "filter_id_type");
 	RNA_def_property_enum_items(prop, rna_enum_id_type_items);
 	RNA_def_property_ui_text(prop, "Filter ID Type", "Data-block type to show");
+	RNA_def_property_translation_context(prop, BLT_I18NCONTEXT_ID_ID);
 }
 
 static void rna_def_space_view3d_shading(BlenderRNA *brna)
@@ -3133,7 +3134,6 @@ static void rna_def_space_view3d(BlenderRNA *brna)
 	RNA_def_property_update(prop, NC_SPACE | ND_SPACE_VIEW3D, NULL);
 
 	prop = RNA_def_property(srna, "clip_start", PROP_FLOAT, PROP_DISTANCE);
-	RNA_def_property_float_sdna(prop, NULL, "near");
 	RNA_def_property_range(prop, 1e-6f, FLT_MAX);
 	RNA_def_property_ui_range(prop, 0.001f, FLT_MAX, 10, 3);
 	RNA_def_property_float_default(prop, 0.1f);
@@ -3141,7 +3141,6 @@ static void rna_def_space_view3d(BlenderRNA *brna)
 	RNA_def_property_update(prop, NC_SPACE | ND_SPACE_VIEW3D, NULL);
 
 	prop = RNA_def_property(srna, "clip_end", PROP_FLOAT, PROP_DISTANCE);
-	RNA_def_property_float_sdna(prop, NULL, "far");
 	RNA_def_property_range(prop, 1e-6f, FLT_MAX);
 	RNA_def_property_ui_range(prop, 0.001f, FLT_MAX, 10, 3);
 	RNA_def_property_float_default(prop, 1000.0f);
@@ -3309,7 +3308,7 @@ static void rna_def_space_view3d(BlenderRNA *brna)
 				RNA_def_property_boolean_negative_sdna(
 				        prop, NULL, view_mask_member[mask_index], info[type_index].type_mask);
 				RNA_def_property_ui_text(prop, info[type_index].name, "");
-				RNA_def_property_update(prop, NC_SPACE | ND_SPACE_SEQUENCER, NULL);
+				RNA_def_property_update(prop, NC_SPACE | ND_SPACE_VIEW3D, NULL);
 
 			}
 		}
@@ -3450,7 +3449,7 @@ static void rna_def_space_buttons(BlenderRNA *brna)
 	PropertyRNA *prop;
 
 	srna = RNA_def_struct(brna, "SpaceProperties", "Space");
-	RNA_def_struct_sdna(srna, "SpaceButs");
+	RNA_def_struct_sdna(srna, "SpaceProperties");
 	RNA_def_struct_ui_text(srna, "Properties Space", "Properties space data");
 
 	prop = RNA_def_property(srna, "context", PROP_ENUM, PROP_NONE);
@@ -4090,7 +4089,7 @@ static void rna_def_space_graph(BlenderRNA *brna)
 
 
 	srna = RNA_def_struct(brna, "SpaceGraphEditor", "Space");
-	RNA_def_struct_sdna(srna, "SpaceIpo");
+	RNA_def_struct_sdna(srna, "SpaceGraph");
 	RNA_def_struct_ui_text(srna, "Space Graph Editor", "Graph Editor space data");
 
 	/* mode */
@@ -4360,7 +4359,7 @@ static void rna_def_fileselect_params(BlenderRNA *brna)
 	static const EnumPropertyItem display_size_items[] = {
 	    {32,    "TINY",     0,      "Tiny", ""},
 	    {64,    "SMALL",    0,      "Small", ""},
-	    {128,   "NORMAL",   0,      "Normal", ""},
+	    {128,   "NORMAL",   0,      "Regular", ""},
 	    {256,   "LARGE",    0,      "Large", ""},
 	    {0, NULL, 0, NULL, NULL},
 	};