+++ release/scripts/addons/blender_id/CHANGELOG.md
@@ -6,6 +6,8 @@
 - API change: `blender_id.get_subclient_user_id()` now returns `''` instead of `None` when the user
   is not logged in.
 - Log which Blender ID instance is communicated with.
+- Show which Blender ID instance is communicated with in the addon preferences,
+  if it was overridden by setting the BLENDER_ID_ENDPOINT environment variable.
 
 
 # Version 1.5 (released 2018-07-03)